renamed the components to TextOptionViewCover/Uncover
